pdf-canvas: move remaining canvas tests to integration suite
Move graphics-state and recording-canvas tests out of source modules
and exercise their behavior through RecordingCanvas and public APIs.

Extend the shared replay observer with mask-layer tracking so soft-mask
coverage no longer needs a dedicated backend implementation.
