+++ crates/pdf-canvas/src/canvas_graphics_state_ops.rs
@@ -194,217 +194,3 @@ impl<B: CanvasBackend> GraphicsStateOps for PdfCanvas<'_, B> {
         Ok(())
     }
 }
-
-#[cfg(test)]
-#[allow(clippy::expect_used)]
-mod tests {
-    use std::{collections::HashMap, rc::Rc};
-
-    use pdf_content_stream::ContentStream;
-    use pdf_document::page::PdfPage;
-    use pdf_graphics::{MaskMode, PathFillType, color::Color, rect::Rect};
-    use pdf_resources::{
-        external_graphics_state::{ExternalGraphicsState, ExternalGraphicsStateKey, SoftMask},
-        form::FormXObject,
-        resource::Resource,
-        resources::Resources,
-        xobject::XObject,
-    };
-
-    use crate::canvas_backend::{Image, Shader};
-
-    use super::*;
-
-    #[derive(Default)]
-    struct MaskCountingCanvas {
-        begin_mask_count: usize,
-    }
-
-    impl CanvasBackend for MaskCountingCanvas {
-        fn fill_path(
-            &mut self,
-            _path: &pdf_graphics::pdf_path::PdfPath,
-            _fill_type: PathFillType,
-            _color: Color,
-            _shader: &Option<Shader>,
-            _blend_mode: Option<pdf_graphics::BlendMode>,
-        ) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-
-        fn stroke_path(
-            &mut self,
-            _path: &pdf_graphics::pdf_path::PdfPath,
-            _color: Color,
-            _line_width: f32,
-            _stroke_style: &crate::stroke_style::StrokeStyle,
-            _shader: &Option<Shader>,
-            _blend_mode: Option<pdf_graphics::BlendMode>,
-        ) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-
-        fn set_clip_region(
-            &mut self,
-            _path: &pdf_graphics::pdf_path::PdfPath,
-            _mode: PathFillType,
-        ) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-
-        fn width(&self) -> f32 {
-            100.0
-        }
-
-        fn height(&self) -> f32 {
-            100.0
-        }
-
-        fn save(&mut self) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-
-        fn restore(&mut self) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-
-        fn draw_image_rect(
-            &mut self,
-            _image: &Image<'_>,
-            _blend_mode: Option<pdf_graphics::BlendMode>,
-            _dest_rect: Rect,
-            _image_rotation: Option<f32>,
-        ) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-
-        fn draw_inline_image(
-            &mut self,
-            _image: &Image<'_>,
-            _blend_mode: Option<pdf_graphics::BlendMode>,
-            _dest_rect: Rect,
-            _image_rotation: Option<f32>,
-        ) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-
-        fn begin_mask_layer(
-            &mut self,
-            _mask: &Arc<RecordingCanvas>,
-            _transform: &Transform,
-            _mask_mode: MaskMode,
-        ) -> Result<(), PdfCanvasError> {
-            self.begin_mask_count += 1;
-            Ok(())
-        }
-
-        fn end_mask_layer(
-            &mut self,
-            _mask: &Arc<RecordingCanvas>,
-            _transform: &Transform,
-            _mask_mode: MaskMode,
-        ) -> Result<(), PdfCanvasError> {
-            Ok(())
-        }
-    }
-
-    fn page() -> PdfPage {
-        PdfPage {
-            contents: None,
-            annotations: None,
-            media_box: None,
-            resources: None,
-            annotation_id_high_watermark: 0,
-        }
-    }
-
-    fn soft_mask_form_resource(bbox: Rect) -> Resources {
-        let form = FormXObject {
-            bbox,
-            matrix: None,
-            resources: None,
-            content_stream: ContentStream {
-                operators: Vec::new(),
-                id: 1,
-            },
-        };
-        let state = ExternalGraphicsState {
-            params: vec![ExternalGraphicsStateKey::SoftMask(Some(Box::new(
-                SoftMask {
-                    mask_type: MaskMode::Alpha,
-                    shape: XObject::Form(Box::new(form)),
-                },
-            )))],
-        };
-
-        Resources {
-            ext_g_states: HashMap::from([(
-                "GS0".to_string(),
-                Resource::ExternalGraphicsState(Rc::new(state)),
-            )]),
-            ..Default::default()
-        }
-    }
-
-    fn dash_pattern_resource() -> Resources {
-        let state = ExternalGraphicsState {
-            params: vec![ExternalGraphicsStateKey::DashPattern(
-                DashPattern::new(&[3.0, 1.0], 2.0)
-                    .expect("dash pattern should be valid")
-                    .expect("dash pattern should be present"),
-            )],
-        };
-
-        Resources {
-            ext_g_states: HashMap::from([(
-                "GS0".to_string(),
-                Resource::ExternalGraphicsState(Rc::new(state)),
-            )]),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn external_graphics_state_dash_pattern_updates_current_state() {
-        let page = page();
-        let resources = dash_pattern_resource();
-        let mut backend = MaskCountingCanvas::default();
-        let mut canvas = PdfCanvas::new(&mut backend, &page, None).expect("canvas should build");
-        canvas.current_state_mut().expect("state").resources = Some(&resources);
-
-        canvas
-            .set_graphics_state_from_dict("GS0")
-            .expect("dash pattern should be supported");
-
-        let dash_pattern = canvas
-            .current_state()
-            .expect("state")
-            .dash_pattern
-            .as_ref()
-            .expect("dash pattern should be set");
-        assert_eq!(dash_pattern.intervals, vec![3.0, 1.0]);
-        assert_eq!(dash_pattern.phase, 2.0);
-    }
-
-    #[test]
-    fn soft_mask_form_with_zero_area_bbox_is_ignored() {
-        let page = page();
-        let resources = soft_mask_form_resource(Rect {
-            left: 5.0,
-            top: 10.0,
-            right: 5.0,
-            bottom: 20.0,
-        });
-        let mut backend = MaskCountingCanvas::default();
-        let mut canvas = PdfCanvas::new(&mut backend, &page, None).expect("canvas should build");
-        canvas.current_state_mut().expect("state").resources = Some(&resources);
-
-        canvas
-            .set_graphics_state_from_dict("GS0")
-            .expect("zero-area soft mask should be ignored");
-
-        assert!(canvas.mask.is_none());
-        drop(canvas);
-        assert_eq!(backend.begin_mask_count, 0);
-    }
-}
